third_party/blink/web_tests/platform/mac/external/wpt/html/semantics/forms/textfieldselection/selection-not-application-textarea-expected.txt

This is a testharness.js-based test.
FAIL text field selection for the input textarea assert_equals: final selectionDirection expected "forward" but got "none"
Harness: the test ran to completion.
